Fixes and optimizations
- Fixed a bunch of typos.
- Added holdaction monitor.
This is a try of a workaround for the action bu.
For some reason the action seems to bug out (condition overflow?!)
sometimes so the player cant scavenge anymore for a few seconds.
The monitor removes all actions from the player if he is in a action
event and adds them back if the action event is done but this is not
working so far.